RastimWink/requirements.txt
fastapi==0.120
uvicorn==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose=3.5.0
passlib==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.12.3
alembic==1.17.0
email-validator==2.3.0